fix: update type definitions with local Chrysalis

+++ src/Argus.Sync/Extensions/Chrysalis/TransactionInputExtension.cs
@@ -1,5 +1,6 @@
 using Argus.Sync.Data.Models;
-using Chrysalis.Cardano.Models.Core.Transaction;
+using Chrysalis.Cardano.Models.Core.Block.Transaction;
+using Chrysalis.Cardano.Models.Core.Block.Transaction.Output;
 
 namespace Argus.Sync.Extensions.Chrysalis;
 
@@ -8,7 +9,7 @@ public static class TransactionInputExtension
     public static string TransacationId(this TransactionInput transactionInput)
         => Convert.ToHexString(transactionInput.TransactionId.Value).ToLowerInvariant();
 
-    public static int Index(this TransactionInput transactionInput)
+    public static ulong Index(this TransactionInput transactionInput)
         => transactionInput.Index.Value;
 
     public static byte[]? ScriptRef(this TransactionOutput transactionOutput)
